fix(aicode-utils): consolidate git utilities and prevent command injection
- Move git.ts from apps/aicode-toolkit and packages/scaffold-mcp to packages/aicode-utils
- Add '--' separator in git commands to prevent second-order command injection
- Add proper type guards (isExecaError, isGitHubContentItem)
- Update imports across codebase to use @agiflowai/aicode-utils directly
- Fix test files to use consolidated module imports
- Rename test file to follow kebab-case convention
