+++ apps/aicode-toolkit/src/utils/index.ts
@@ -4,13 +4,3 @@
 
 // Banner utilities
 export { displayBanner, displayCompactBanner } from './banner';
-
-// Git utilities
-export {
-  cloneRepository,
-  cloneSubdirectory,
-  fetchGitHubDirectoryContents,
-  findWorkspaceRoot,
-  gitInit,
-  parseGitHubUrl,
-} from './git';
