+++ packages/aicode-utils/package.json
@@ -36,6 +36,7 @@
   },
   "dependencies": {
     "chalk": "5.6.2",
+    "execa": "^9.5.2",
     "js-yaml": "4.1.1",
     "ora": "^9.0.0",
     "pino": "^10.0.0"
