+++ packages/scaffold-mcp/src/utils/index.ts
@@ -2,12 +2,4 @@
  * Utils Barrel Exports
  */
 
-export {
-  cloneRepository,
-  cloneSubdirectory,
-  fetchGitHubDirectoryContents,
-  findWorkspaceRoot,
-  parseGitHubUrl,
-} from './git';
-
 export { PaginationHelper, type PaginationResult } from './pagination';
